UPSTREAM: drivers: base: add coredump driver ops
This adds the coredump driver operation. When the driver defines it
a coredump file is added in the sysfs folder of the device upon
driver binding. The file is removed when the driver is unbound.
User-space can trigger a coredump for this device by echo'ing to
the coredump file.

Signed-off-by: Arend van Spriel <aspriel@gmail.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 3c47d19ff4dccf1500c33bcbe3b5bc804907a0da)

BUG=None
TEST=Built

Signed-off-by: Manish Mandlik <mmandlik@google.com>
Change-Id: I930f1a18c9f09b2ea1d756d97afc40c3e1cd1316
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3830268
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Tested-by: Manish Mandlik <mmandlik@chromium.org>
Commit-Queue: Manish Mandlik <mmandlik@chromium.org>