CHROMIUM: iwl7000: mac80211: drop A-MSDUs on old ciphers
With old ciphers (WEP and TKIP) we shouldn't be using A-MSDUs
since A-MSDUs are only supported if we know that they are, and
the only practical way for that is HT support which doesn't
support old ciphers.

However, we would normally accept them anyway. Since we check
the MMIC before deaggregating A-MSDUs, and the A-MSDU bit in
the QoS header is not protected in TKIP (or WEP), this enables
attacks similar to CVE-2020-24588. To prevent that, drop A-MSDUs
completely with old ciphers.

BUG=b:187853103, b:187853926
TEST=Verified wifi_matfunc/perf

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 328b8ff10f9dcb9bc08aa069744073a4de3a309a
Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I548e6e71f1ee9cad4b9a37bf212ae7db723587aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2888689
Reviewed-by: Brian Norris <briannorris@chromium.org>