BACKPORT: nl80211: reduce non-split wiphy dump size
When wiphy dumps cannot be split, such as in events or with
older userspace that doesn't support it, the size can today
be too big.

Reduce it, by doing two things:

 1) remove data that couldn't have been present before the
    split capability was introduced since it's new, such as
    HE capabilities

 2) as suggested by Martin Willi, remove management frame
    subtypes from the split dumps, as just (1) isn't even
    enough due to other new code capabilities. This is fine
    as old consumers (really just wpa_supplicant) didn't
    check this data before they got support for split dumps.

Reported-by: Martin Willi <martin@strongswan.org>
Suggested-by: Martin Willi <martin@strongswan.org>
Tested-by: Martin Willi <martin@strongswan.org>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Link: https://lore.kernel.org/r/20200928130655.53bce7873164.I71f06c9a221cd0630429a1a56eeae68a13beca61@changeid
(cherry picked from commit f8d504caa9739f74c7c399e158a7dce8478b866e)

Backport notes:
This change adds a boolean value "large" to indicate that an
nl80211 message is larger than a certain threshold. This
boolean value is used in several places, two of which are
related to features or operations which are not yet
implemented in this version of the kernel. These are
EDMG channels and channel frequency offsets. Because
these features do not exist in 4.19, and adding them in
would require pulling a large number of patches, it seems
safer to just remove those two callsites from this patch.

BUG=b:266832532
TEST=deploy to hatch device. Ensure that NL80211_CMD_NEW_WIPHY
are broadcast as expected and do basic test for WiFi
funtionality.
Signed-off-by: Kevin Lund <kglund@google.com>

Change-Id: Ia11087c968374fac9f532b99f65d39669e551f4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4221345
Reviewed-by: Johannes Berg <johannes.berg@intel.corp-partner.google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>