UPSTREAM: kconfig: qconf: drop Qt4 support
It is possible to keep this compatible with both Qt4 and Qt5, but it is
questionable if it is worth the efforts; it would require us to test
this on both of them, and prevent us from using new features in Qt5.

Qt5 was released in 2012, and now widely available.

Drop the Qt4 support.

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
(cherry picked from commit 7b675649be2217786847dee13597a0ab8502cc40)

BUG=b:242048430
TEST=None

Change-Id: Ica6a5b6dede67923cd5e3c8a0f27f61b38297faa
Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4545590
Reviewed-by: Ross Zwisler <zwisler@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>