CHROMIUM: net: qualcomm: drop remove() and rmmod support
In normal usage, the device and driver remove paths are not exercised.
It also happens they have many bugs.

For some use cases (say, check_ethernet.hook in a test lab), we *do*
exercise device unbind and/or rmmod as a blunt recovery method. In
practice, that just leads to kernel panics or lockups.

We're better off if we simply drop support for remove(), so such scripts
can skip over this method and try something else, like clean reboot.

BUG=b:177380545
TEST=`rmmod essedma` gives EBUSY
TEST=/sys/class/net/xxx/device/driver/unbind no longer exists for eth
     ports

Signed-off-by: Brian Norris <briannorris@chromium.org>
Change-Id: I9b258f0452b84c7b2b86dee84287fe75faf1effa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2689072
Reviewed-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Kan Yan <kyan@chromium.org>