CHROMIUM: ath10k: Validate first subframe of amsdu before processing the list
Validate amsdu list by checking if the first frame has a valid
DA. In certain scenarios a proper msdu can be received as an amsdu when
the A-MSDU present bit gets flipped.
Since this bit is unauthenticated the HW crypto engine can pass the frame
to driver without any error indication. This could result in processing
unintended subframes collected in the amsdu list.

Comparing the non-aggregated msdu and an amsdu, the fields of the first
subframe DA matches with the llc/snap header fields of normal msdu.
In order to avoid processing such frames, add a validation to
filter such amsdu frames where the first subframe header DA matches
with the llc/snap header pattern.

BUG=b:179418147
TEST=manual testing on G

Change-Id: Ibcaee28f2bf53bb830a35737dd01d3a2ea5d4c08
Signed-off-by: Maharaja Kennadyrajan <mkenna@codeaurora.org>
Signed-off-by: Vadiraj Kakhandaki <vadirajk@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2914541
Tested-by: ANKIT DARSI AMARNATHAGUPTHA <dankit@google.com>
Reviewed-by: Julan Hsu <julanhsu@google.com>