CHROMIUM: mac80211: assure all fragments are encrypted
Do not mix plaintext and encrypted fragments in protected Wi-Fi
networks. This fixes CVE-2020-26147.

Previously, an attacker was able to first forward a legitimate encrypted
fragment towards a victim, followed by a plaintext fragment. The
encrypted and plaintext fragment would then be reassembled. For further
details see Section 6.3 and Appendix D in the paper "Fragment and Forge:
Breaking Wi-Fi Through Frame Aggregation and Fragmentation".

Because of this change there are now two equivalent conditions in the
code to determine if a received fragment requires sequential PNs, so we
also move this test to a separate function to make the code easier to
maintain.

BUG=b:179418147
TEST=manual testing on G

Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@kuleuven.be>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: vadiraj kakhandaki <vadirajk@google.com>
Change-Id: I5dc3af7b41307fe138ba9db476af8441c1c96860
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2914549
Reviewed-by: Maharaja Kennadyrajan <mkenna@qualcomm.corp-partner.google.com>
Reviewed-by: Julan Hsu <julanhsu@google.com>