CHROMIUM: mac80211: prevent mixed key and fragment cache attacks
Simultaneously prevent mixed key attacks (CVE-2020-24587) and fragment
cache attacks (CVE-2020-24586). This is accomplished by assigning a
unique color to every key (per interface) and using this to track which
key was used to decrypt a fragment. When reassembling frames, it is
now checked whether all fragments were decrypted using the same key.

To assure that fragment cache attacks are also prevented, the ID that is
assigned to keys is unique even over (re)associations and (re)connects.
This means fragments separated by a (re)association or (re)connect will
not be reassembled. Because mac80211 now also prevents the reassembly of
mixed encrypted and plaintext fragments, all cache attacks are prevented.

BUG=b:179418147
TEST=manual testing on G

Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@kuleuven.be>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: vadiraj kakhandaki <vadirajk@google.com>
Change-Id: Ibfd887fd3a269d6bdd3343a1c6492ca111b9c1a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2914913
Reviewed-by: Maharaja Kennadyrajan <mkenna@qualcomm.corp-partner.google.com>
Reviewed-by: Julan Hsu <julanhsu@google.com>