UPSTREAM: tracing: Use a no_filter_buffering_ref to stop using the filter buffer
Currently, the trace histograms relies on it using absolute time stamps to
trigger the tracing to not use the temp buffer if filters are set. That's
because the histograms need the full timestamp that is saved in the ring
buffer. That is no longer the case, as the ring_buffer_event_time_stamp()
can now return the time stamp for all events without all triggering a full
absolute time stamp.

Now that the absolute time stamp is an unrelated dependency to not using
the filters. There's nothing about having absolute timestamps to keep from
using the filter buffer. Instead, change the interface to explicitly state
to disable filter buffering that the histogram logic can use.

Link: https://lkml.kernel.org/r/20210316164113.847886563@goodmis.org

Reviewed-by: Tom Zanussi <zanussi@kernel.org>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
(cherry picked from commit b94bc80df64823e676b506f8de7dcf6a688d681e)

BUG=b:316092473
TEST=Run timerlat tracer with config modification

Change-Id: I06409e96279dad7aeda1454177fd90fbb726a0b1
Signed-off-by: Baili Deng <bailideng@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5184619
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>