UPSTREAM: tracing: Add "last_func_repeats" to struct trace_array
The field is used to keep track of the consecutive (on the same CPU) calls
of a single function. This information is needed in order to consolidate
the function tracing record in the cases when a single function is called
number of times.

Link: https://lkml.kernel.org/r/20210415181854.147448-4-y.karadz@gmail.com

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
(cherry picked from commit 20344c54d1c7ab7428e312bbe9b0097750875002)

BUG=b:316092473
TEST=Run timerlat tracer with config modification

Change-Id: If2a128aab0a746e20e2133c3610f9518464bf3e3
Signed-off-by: Baili Deng <bailideng@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5184625
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>