CHROMIUM: mac80211: add fragment cache to sta_info
Prior patches protected against fragmentation cache attacks
by coloring keys, but this shows that it can lead to issues
when multiple stations use the same sequence number. Add a
fragment cache to struct sta_info (in addition to the one in
the interface) to separate fragments for different stations
properly.

This then automatically clear most of the fragment cache when a
station disconnects (or reassociates) from an AP, or when client
interfaces disconnect from the network, etc.

On the way, also fix the comment there since this brings us in line
with the recommendation in 802.11-2016 ("An AP should support ...").
Additionally, remove a useless condition (since there's no problem
purging an already empty list).

BUG=b:179418147
TEST=manual testing on G

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: vadiraj kakhandaki <vadirajk@google.com>
Change-Id: I1cc0ff3396e59e367ac1b304d342991bc0b01671
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2914918
Reviewed-by: Maharaja Kennadyrajan <mkenna@qualcomm.corp-partner.google.com>
Reviewed-by: Julan Hsu <julanhsu@google.com>