CHROMIUM: config: Enable CONFIG_NETFILTER_XT_MATCH_STATISTIC
When running DNS proxy (go/cros-dns-proxy), Chrome traffic
must not be redirected to the proxy. In particular, since
it will always share the same Secure DNS settings as the
OS, there is nothing to be gained by having it's DNS traffic
proxied. Any DNS query arriving from Chrome must mean that
one of the following is true:
- A secure hostname is being resolved,
- Secure DNS is disabled,
- A secure query was already attempted and fell back to
  standard resolution.
So, rather than introducing special handling in the DNS
proxy service itself, we can instead handle this directly
in iptables, and bypass the service for Chrome DNS.

Because there can be multiple nameservers at a time, it is
necessary to find a solution that is able to query not only
the first nameserver. With iptable, this can be done using
the statistic module. By using the module, the traffic can
redirected in round robin fashion. This configuration is
available from kernel 2.6.

BUG=b:171363265
TEST=iptables -t nat -I OUTPUT -m statistic --mode nth \
     --every 2 --packet 1 -j ACCEPT;

Signed-off-by: Jason Jeremy Iman <jasongustaman@google.com>
Change-Id: I63c4d2a8c51c0aa7c54210be1cc4e66f3ee2fd0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2885400
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
(cherry picked from commit 46cdfaee0fe42172c936f6d44991f67ea000e2ca)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2927949