UPSTREAM: tracing: Add better comments for the filtering temp buffer use case
When filtering is enabled, the event is copied into a temp buffer instead
of being written into the ring buffer directly, because the discarding of
events from the ring buffer is very expensive, and doing the extra copy is
much faster than having to discard most of the time.

As that logic is subtle, add comments to explain in more detail to what is
going on and how it works.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
(cherry picked from commit 8f0901cda14d3be38cd2196d8cf61cdf3b368e34)

BUG=b:316092473
TEST=Run timerlat tracer with config modification

Change-Id: I5c847b3533a8c3348de451af707814b43a89d139
Signed-off-by: Baili Deng <bailideng@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5184621
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>