UPSTREAM: drm/i915: Rename i915_gem_context_get_vm_rcu to i915_gem_context_get_eb_vm
The important part isn't so much that this does an rcu lookup - that's
more an implementation detail, which will also be removed.

The thing that makes this different from other functions is that it's
gettting you the vm that batchbuffers will run in for that gem
context, which is either a full ppgtt stored in gem->ctx, or the ggtt.

We'll make more use of this function later on.

Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>
Cc: Jon Bloomfield <jon.bloomfield@intel.com>
Cc: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: "Thomas Hellström" <thomas.hellstrom@linux.intel.com>
Cc: Matthew Auld <matthew.auld@intel.com>
Cc: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Cc: Dave Airlie <airlied@redhat.com>
Cc: Jason Ekstrand <jason@jlekstrand.net>
Link: https://patchwork.freedesktop.org/patch/msgid/20210902142057.929669-5-daniel.vetter@ffwll.ch
(cherry picked from commit c6d04e48d2e6d0e41c4cc4098c5494713086b597)

BUG=b:278110760
TEST=emerge-volteer sys-kernel/chromeos-kernel-5_10

Change-Id: I1ab6f070e88247d57e506ac3648ea77e19b24f6e
Signed-off-by: Gareth Yu <gareth.yu@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4586859
Reviewed-by: Chia-I Wu <olv@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Matt Turner <msturner@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Matt Turner <msturner@google.com>
Reviewed-by: Matt Turner <msturner@google.com>