CHROMIUM: drm/i915/pxp: handle init/term race condition
If the PXP TEE message to initialize a session is interrupted by a
signal, instead of waiting for the response to confirm the session has
been initialized the driver exits immediately. The message however was
still sent and the session ends up being created later.

Normally upon app crash, intel_pxp_file_close() will cleanup and
terminate any associated sessions.
In this case however, a race condition occurs since the terminate is
called after the signal but before the session has finished initializing
so the terminate does nothing and afterwards the session is created
without the driver knowing.
Eg. SW thinks the sessions is free but HW session is initialized

To fix this, if we find a session where the SW and HW state don't align,
attempt to terminate the session before trying to reserve it again.
its possible that SW state of pxp->reserved_sessions maybe out of sync with HW.
So lets combine active_hw_slots in for termination which would
normally match pxp->reserved_sessions

Signed-off-by: Juston Li <juston.li@intel.com>
Reviewed-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Signed-off-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>

Fixes: bc28dfd963c6 ("CHROMIUM: drm/i915/pxp: handle init/term race condition")

BUG=b:257876317 b:284492421
TEST=None

Change-Id: Ia5c16e8b2c1fed7ba9838ac18d76b0fa470976f9
Signed-off-by: Narasimha Rao Rayala <narasimha.rao.rayala@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4872417
Reviewed-by: Paz Zcharya <pazz@google.com>
Reviewed-by: Juston Li <justonli@google.com>