UPSTREAM: HID: apple: Report Magic Keyboard battery over USB
When connected over USB, the Apple Magic Keyboard 2015 registers 3
different interfaces. One of them is used to report the battery level.

However, unlike when connected over Bluetooth, the battery level is not
reported automatically and it is required to fetch it manually.

Add a new quirk to fix the battery report descriptor and a timer to
fetch the battery level.

Signed-off-by: José Expósito <jose.exposito89@gmail.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 6e143293e17a73c9313f91c5ca3aaacbaef030cf)

BUG=b:357788547, b:357786089
TEST=Build and flash kernel on blacktip, verify that top row keys work as
     expected

Change-Id: I2a0ad908f373575d659f2f52a269497fba4bd133
Signed-off-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6003567
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>