UPSTREAM: r8152: sync ocp base
There are some chances that the actual base of hardware is different
from the value recorded by driver, so we have to reset the variable
of ocp_base to sync it.

Set ocp_base to -1. Then, it would be updated and the new base would be
set to the hardware next time.

Signed-off-by: Hayes Wang <hayeswang@realtek.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit b24edca309535c2d9af86aab95d64065f6ef1d26)

BUG=none
TEST=test_that -b brya  192.168.86.163 network_EthernetStressPlug --args num_iterations=1000
     using redrix (brya) + ServoV4.0 connected to netgear GS108 switch

Change-Id: Id5214d0f539ef518ee675fa5b64ae81b0882b2af
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4129444
Reviewed-by: Ying Hsu <yinghsu@chromium.org>
Reviewed-by: Hayes Wang <hayeswang@realtek.com>
Reviewed-by: Sean Paul <sean@poorly.run>