UPSTREAM: drm: Plumb debugfs_init through to panels
We'd like panels to be able to add things to debugfs underneath the
connector's directory. Let's plumb it through. A panel will be able to
put things in a "panel" directory under the connector's
directory. Note that debugfs is not ABI and so it's always possible
that the location that the panel gets for its debugfs could change in
the future.

NOTE: this currently only works if you're using a modern
architecture. Specifically the plumbing relies on _both_
drm_bridge_connector and drm_panel_bridge. If you're not using one or
both of these things then things won't be plumbed through.

As a side effect of this change, drm_bridges can also get callbacks to
put stuff underneath the connector's debugfs directory. At the moment
all bridges in the chain have their debugfs_init() called with the
connector's root directory.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Javier Martinez Canillas <javierm@redhat.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20220204161245.v2.2.Ib0bd5346135cbb0b63006b69b61d4c8af6484740@changeid
(cherry picked from commit 2509969a9862b522d2208e8663057fb227556687)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:282117958
TEST= IGT tool can read output_bpc properly.

Change-Id: I40335ff6f7d4e700ba0852dcb08835ab9ae8b1c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4525938
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Shawn C Lee <shawn.c.lee@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Shawn C Lee <shawn.c.lee@intel.com>