UPSTREAM: HID: apple: Report Magic Keyboard 2021 with fingerprint reader battery over USB
Like the Apple Magic Keyboard 2015, when connected over USB, the 2021
version with fingerprint reader registers 2 different interfaces. One of
them is used to report the battery level.

However, unlike when connected over Bluetooth, the battery level is not
reported automatically and it is required to fetch it manually.

Add the APPLE_RDESC_BATTERY quirk to fix the battery report descriptor
and manually fetch the battery level.

Tested with the ANSI variant of the keyboard with and without numpad.

Signed-off-by: José Expósito <jose.exposito89@gmail.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit cbfcfbfc384890a062a5d0cc4792df094a6cc7a8)

BUG=b:357788547, b:357786089
TEST=Build and flash kernel on blacktip, verify that top row keys work as
     expected

Change-Id: I4f9ceb3ded8af4b50bd0a4ea2793964570eb849a
Signed-off-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6003576
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>