CHROMIUM: iwl7000: mac80211: fix and clean up puncturing support
We must not unconditionally access eht_oper->disable_subchannel_bitmap
down in ieee80211_get_disabled_subchannel_bitmap(), since that will be
called even if there was no bitmap present in the EHT Operation Element
in the first place.

Fix that, but also be more precise about the naming (bitmap is the one
from the element, extracted is the extracted one), and pass around the
sdata pointer to be easier to read.

Change-Id: Ifcf549decf466389595644a425176726103f1829
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 02a339c1bf65cba942fcf60cac56002d96d97703