UPSTREAM: tracing/osnoise: Make osnoise_main to sleep for microseconds
osnoise's runtime and period are in the microseconds scale, but it is
currently sleeping in the millisecond's scale. This behavior roots in the
usage of hwlat as the skeleton for osnoise.

Make osnoise to sleep in the microseconds scale. Also, move the sleep to
a specialized function.

Link: https://lkml.kernel.org/r/302aa6c7bdf2d131719b22901905e9da122a11b2.1645197336.git.bristot@kernel.org

Cc: Ingo Molnar <mingo@redhat.com>
Signed-off-by: Daniel Bristot de Oliveira <bristot@kernel.org>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit dd990352f01ee9a6c6eee152e5d11c021caccfe4)

BUG=b:316092473
TEST=Run timerlat tracer with config modification

Change-Id: I4ad21a350c04b9e5f0a2d9915757b251e2928597
Signed-off-by: Baili Deng <bailideng@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5237659
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>