UPSTREAM: drm/i915/guc: Do not complain about stale reset notifications
It is possible for reset notifications to arrive for a context that is
in the process of being banned. So don't flag these as an error, just
report it as informational (because it is still useful to know that
resets are happening even if they are being ignored).

v2: Better wording for the message (review feedback from Tvrtko).
v3: Fix rebase issue (review feedback from Daniele).

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20220225015232.1939497-1-John.C.Harrison@Intel.com
(cherry picked from commit e2a1e7abaee128020fde54d386ec8959b1e9eb61)

BUG=b:278110760
TEST=emerge-volteer sys-kernel/chromeos-kernel-5_10

Change-Id: I4114de1b2b002236d5a4d5edf146e870eabec2df
Signed-off-by: Gareth Yu <gareth.yu@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4587682
Reviewed-by: Matt Turner <msturner@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Chia-I Wu <olv@google.com>
Tested-by: Matt Turner <msturner@google.com>
Commit-Queue: Matt Turner <msturner@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>