UPSTREAM: drm/i915/guc: Better name for context id limit
The LRC descriptor pool is going away. So, stop using it as the limit
for how many context ids are available. Instead, size the pool
according to the number of contexts allowed. Note that this is just a
naming change, the actual limit is identical in value.

While at it, also update a kzalloc(sizeof()*count) to be a
kcalloc(count,size).

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20220302003357.4188363-4-John.C.Harrison@Intel.com
(cherry picked from commit d124902242241171b704d051ef2830dd344dff84)

BUG=b:278110760
TEST=emerge-volteer sys-kernel/chromeos-kernel-5_10

Change-Id: I4101e19d85cb5d37546e44fa3e02504a16b5baff
Signed-off-by: Gareth Yu <gareth.yu@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4587684
Reviewed-by: Chia-I Wu <olv@google.com>
Commit-Queue: Matt Turner <msturner@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Matt Turner <msturner@google.com>
Reviewed-by: Matt Turner <msturner@google.com>