BACKPORT: platform/chrome: Add ChromeOS ACPI device driver
The x86 Chromebooks have the ChromeOS ACPI device. This driver attaches
to the ChromeOS ACPI device and exports the values reported by ACPI in a
sysfs directory. This data isn't present in ACPI tables when read
through ACPI tools, hence a driver is needed to do it. The driver gets
data from firmware using the ACPI component of the kernel. The ACPI values
are presented in string form (numbers as decimal values) or binary
blobs, and can be accessed as the contents of the appropriate read only
files in the standard ACPI device's sysfs directory tree. This data is
consumed by the ChromeOS user space.

Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Andy Shevchenko <andy.shevchenko@gmail.com>
Reviewed-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Acked-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Co-developed-by: Muhammad Usama Anjum <usama.anjum@collabora.com>
Signed-off-by: Muhammad Usama Anjum <usama.anjum@collabora.com>
Signed-off-by: Tzung-Bi Shih <tzungbi@kernel.org>
Link: https://lore.kernel.org/r/Yn4OKYrtV35Dv+nd@debian-BULLSEYE-live-builder-AMD64
(cherry picked from commit 0a4cad9c11ad46662ede48d94f08ecb7cd9f6916)

Conflicts:
   Documentation/acpi/chromeos-acpi-device.rst
   Documentation/firmware-guide/acpi/index.rst
[Documentation shouldn't be a concern.  Removed them as "acpi/" and
 "firmware-guide/" aren't in chromeos-4.14.]
   drivers/platform/chrome/Kconfig
   drivers/platform/chrome/Makefile
[Contextual conflicts.]

BUG=b:291036452
TEST=ls /sys/devices/platform/GOOG0016\:00/

Change-Id: I93285dff1aabc03d0e116dea23a29f6de1034f99
Signed-off-by: Tzung-Bi Shih <tzungbi@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4777032
Reviewed-by: Guenter Roeck <groeck@chromium.org>