CHROMIUM: iwl7000: mvm: add primary 80 known for EHT TLV
We calculate based on chandef the place of the control channel inside
the wide band, this is used for understanding the N value in 11be D.5
table 9-53a for column PHY MU/MRU index.

To avoid the need to calculate every frame the value we do it once
monitor vif is added.

Change-Id: I291ee480252d098f62d9ec39040284d3e521d88e
Signed-off-by: Mordechay Goodstein <mordechay.goodstein@intel.com>
iwl7000-tree: 6132db455ba3d10017e2f5eff6770a111d371579