CHROMIUM: iwl7000: mvm: remove setting of 'sta' parameter
cppcheck reports
[drivers/net/wireless/intel/iwlwifi/mvm/rs.c:2686]: (warning) Assignment of function parameter has no effect outside the function. Did you forget dereferencing it?

The setting of the 'sta' parameter is not needed.  In the if-check that sets it
to NULL, mvm_sta is also set to NULL.  Then the next statement checks if
mvm_sta is NULL and does an early return. So remove setting sta.

Change-Id: I2f723f9d44f65720baaf3e84b72109759350a8f5
Signed-off-by: Tom Rix <trix@redhat.com>
iwl7000-tree: bb6e70b72c83022be9e72a287d8b5a1befb0537e