CHROMIUM: iwl7000: mvm: fix the linkId in debugfs beacon injection hooks
When we inject an IE in the beacon we need to assign the linkId only for
the new firmware versions.
Don't feed the linkId to a firmware that is expecting the old
templateId.

Change-Id: Ib33cbc32968efa94929488c137ea4f8d7fe0a233
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: c4d281ae7243f17e1454488599a5bdf33853ac63